Change solution to a "fake" WPF program
