+++ VPinballX.starter.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.7.34024.191
+VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VPinballX.starter", "VPinballX.starter\VPinballX.starter.csproj", "{0E18870A-4A24-4C52-B7B0-E761BADD9D9D}"
 EndProject
@@ -10,14 +10,24 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Projektmappenelemente", "Pr
 		VPinballX.starter\VPinballX.starter.ico = VPinballX.starter\VPinballX.starter.ico
 	EndProjectSection
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{16A264BE-A143-4660-8857-12CAFE240D7C}"
+	ProjectSection(SolutionItems) = preProject
+		VPinballX.starter\App.xaml = VPinballX.starter\App.xaml
+		VPinballX.starter\App.xaml.cs = VPinballX.starter\App.xaml.cs
+		VPinballX.starter\AssemblyInfo.cs = VPinballX.starter\AssemblyInfo.cs
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Release and Deploy|Any CPU = Release and Deploy|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{0E18870A-4A24-4C52-B7B0-E761BADD9D9D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0E18870A-4A24-4C52-B7B0-E761BADD9D9D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0E18870A-4A24-4C52-B7B0-E761BADD9D9D}.Release and Deploy|Any CPU.ActiveCfg = Release|Any CPU
+		{0E18870A-4A24-4C52-B7B0-E761BADD9D9D}.Release and Deploy|Any CPU.Build.0 = Release|Any CPU
 		{0E18870A-4A24-4C52-B7B0-E761BADD9D9D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0E18870A-4A24-4C52-B7B0-E761BADD9D9D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
