+++ VPinballX.starter/App.xaml
@@ -1,0 +1,10 @@
+﻿<Application x:Class="VPinballX.starter.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:VPinballX.starter"
+             Startup="Application_Startup">
+    <Application.Resources>
+        <!--StartupUri="MainWindow.xaml" -->
+
+    </Application.Resources>
+</Application>
